# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /taxes:
    get:
      operationId: listTaxes
      tags: [taxes]
      summary: List all taxes for the current company
      responses:
        "200":
          description: Taxes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tax"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTax
      tags: [taxes]
      summary: Create a tax
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tax"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /taxes/seed-defaults:
    post:
      operationId: seedDefaultTaxes
      tags: [taxes]
      summary: Seed the company's country default taxes (only when none exist)
      responses:
        "201":
          description: Seeded taxes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tax"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Taxes already exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /taxes/{id}:
    patch:
      operationId: updateTax
      tags: [taxes]
      summary: Update a tax
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tax"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTax
      tags: [taxes]
      summary: Delete a tax
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Tax:
      type: object
      required: [id, nameAr, kind, rate, isActive, createdAt]
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        kind:
          type: string
          enum: [vat, wht, income, payroll, zakat]
        rate:
          type: number
        serviceNature:
          type: ["string", "null"]
        linkedAccountId:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    TaxInput:
      type: object
      required: [nameAr, kind, rate]
      properties:
        nameAr:
          type: string
          minLength: 1
        nameEn:
          type: ["string", "null"]
        kind:
          type: string
          enum: [vat, wht, income, payroll, zakat]
        rate:
          type: number
          minimum: 0
        serviceNature:
          type: ["string", "null"]
        linkedAccountId:
          type: ["string", "null"]
          format: uuid
        isActive:
          type: boolean
    TaxUpdate:
      type: object
      properties:
        nameAr:
          type: string
          minLength: 1
        nameEn:
          type: ["string", "null"]
        kind:
          type: string
          enum: [vat, wht, income, payroll, zakat]
        rate:
          type: number
          minimum: 0
        serviceNature:
          type: ["string", "null"]
        linkedAccountId:
          type: ["string", "null"]
          format: uuid
        isActive:
          type: boolean
...